+++ IBTool/changelog.md
@@ -1,3 +1,9 @@
-﻿# Changelog
-## 1.0.0 (2020-10-27)
- - New: First Release
+﻿# Changelog  
+
+## 1.0.9 (2020-10-28)  
+ - New: Added 4 functions  
+ - New: Added display information for Audit logs and Exchange ABP.  
+ - New: Added 3 buttons to call the functions.  
+
+## 1.0.0 (2020-10-27)  
+ - New: Project Start